Broke apart __init__.py into smaller modules
