+++ anvio/workflows/sra_download/Snakefile
@@ -4,6 +4,7 @@ import json
 import anvio
 import argparse
 import pandas as pd
+import anvio.workflows as w
 
 from anvio.errors import ConfigError
 from anvio.workflows.sra_download import SRADownloadWorkflow
